Refactor chat components and update type imports
- Changed import paths for types in chat-related components to use a centralized types file.
- Removed unused chat components (ChannelHeader, ChannelList, MessageArea, MessageInput, MessageItem, and send-message-event).
- Consolidated chat channel entry logic into ChannelContainer component.
- Updated ChatLayout to integrate new components and maintain functionality.
- Ensured all components are using the new types from the centralized types file.
