+++ src/app/sse/_components/channel-header.tsx
@@ -4,7 +4,6 @@ import { useChattingProvider } from "@/app/sse/_components/chatting-provider";
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover";
-import { User } from "./types";
 import { format } from "date-fns";
 import { ko } from "date-fns/locale";
