+++ src/app/sse/_components/channel-list.tsx
