+++ src/app/sse/_components/message-item.tsx
